+++ native/src/main/java/io/ballerina/stdlib/http/api/Resource.java
@@ -1,0 +1,24 @@
+package io.ballerina.stdlib.http.api;
+
+import io.ballerina.runtime.api.types.ResourceMethodType;
+import io.ballerina.stdlib.http.api.service.signature.ParamHandler;
+
+import java.util.List;
+
+/**
+ * Interface for HTTP Resource classes.
+ */
+public interface Resource {
+
+    public List<String> getMethods();
+
+    ParamHandler getParamHandler();
+
+    ResourceMethodType getBalResource();
+
+    boolean isTreatNilableAsOptional();
+
+    String getWildcardToken();
+
+    Service getParentService();
+}
