+++ native/src/main/java/io/ballerina/stdlib/http/api/Service.java
@@ -1,0 +1,15 @@
+package io.ballerina.stdlib.http.api;
+
+import io.ballerina.runtime.api.values.BMap;
+import io.ballerina.runtime.api.values.BString;
+
+/**
+ * Interface for HTTP Service classes.
+ */
+public interface Service {
+    BMap<BString, Object> getCompressionConfig();
+
+    String getChunkingConfig();
+
+    String getMediaTypeSubtypePrefix();
+}
