+++ native/src/main/java/io/ballerina/stdlib/http/api/nativeimpl/ExternRequestContext.java
@@ -28,13 +28,15 @@ import io.ballerina.stdlib.http.api.HttpUtil;
  * Utilities related to HTTP request context.
  */
 public class ExternRequestContext {
-    public static Object next(BObject requestCtx, int interceptorId) {
+    public static Object next(BObject requestCtx) {
         BArray interceptors = (BArray) requestCtx.getNativeData(HttpConstants.HTTP_INTERCEPTORS);
         if (interceptors != null) {
+            int interceptorId = (int) requestCtx.getNativeData(HttpConstants.INTERCEPTOR_SERVICE_INDEX);
             requestCtx.addNativeData(HttpConstants.REQUEST_CONTEXT_NEXT, true);
             if (interceptorId < interceptors.size()) {
                 return interceptors.get(interceptorId);
             }
+            requestCtx.addNativeData(HttpConstants.INTERCEPTOR_SERVICE_INDEX, interceptorId + 1);
             return null;
         } else {
             return HttpUtil.createHttpError("request context object does not contain the configured " +
