+++ native/src/test/java/io/ballerina/stdlib/http/api/HttpServiceTest.java
@@ -31,59 +31,59 @@ import java.util.Arrays;
 import java.util.Collections;
 
 /**
- * A unit test class for http module {@link BaseService} class functions.
+ * A unit test class for http module {@link HttpService} class functions.
  */
-public class BaseServiceTest {
+public class HttpServiceTest {
 
     @Test
     public void testKeepAlive() {
         BObject service = TestUtils.getNewServiceObject("hello");
-        BaseService baseService = new BaseService(service);
-        Assert.assertTrue(baseService.isKeepAlive());
+        HttpService httpService = new HttpService(service);
+        Assert.assertTrue(httpService.isKeepAlive());
 
-        baseService.setKeepAlive(false);
-        Assert.assertFalse(baseService.isKeepAlive());
+        httpService.setKeepAlive(false);
+        Assert.assertFalse(httpService.isKeepAlive());
     }
 
     @Test
     public void testNullServiceBasePath() {
         BObject service = TestUtils.getNewServiceObject("hello");
-        BaseService baseService = new BaseService(service);
-        baseService.setBasePath(null);
+        HttpService httpService = new HttpService(service);
+        httpService.setBasePath(null);
 
-        Assert.assertEquals(baseService.getBasePath(), "/hello");
+        Assert.assertEquals(httpService.getBasePath(), "/hello");
 
         service = TestUtils.getNewServiceObject("$hello");
-        baseService = new BaseService(service);
-        baseService.setBasePath(null);
+        httpService = new HttpService(service);
+        httpService.setBasePath(null);
 
-        Assert.assertEquals(baseService.getBasePath(), "/");
+        Assert.assertEquals(httpService.getBasePath(), "/");
     }
 
     @Test
     public void testEmptyNullServiceBasePath() {
         BObject service = TestUtils.getNewServiceObject("hello");
-        BaseService baseService = new BaseService(service);
-        baseService.setBasePath(" ");
+        HttpService httpService = new HttpService(service);
+        httpService.setBasePath(" ");
 
-        Assert.assertEquals(baseService.getBasePath(), "/hello");
+        Assert.assertEquals(httpService.getBasePath(), "/hello");
     }
 
     @Test
     public void testNotNullServiceBasePath() {
         BObject service = TestUtils.getNewServiceObject("hello");
-        BaseService baseService = new BaseService(service);
-        baseService.setBasePath("ballerina");
+        HttpService httpService = new HttpService(service);
+        httpService.setBasePath("ballerina");
 
-        Assert.assertEquals(baseService.getBasePath(), "/ballerina");
+        Assert.assertEquals(httpService.getBasePath(), "/ballerina");
     }
 
     @Test
     public void testGetPayloadFunctionOfIntrospectionResource() {
         BObject service = TestUtils.getNewServiceObject("hello");
-        BaseService baseService = new BaseService(service);
+        HttpService httpService = new HttpService(service);
         String filePath = "resources/ballerina/http/testopenapidoc.json";
-        HttpIntrospectionResource introspectionResource = new HttpIntrospectionResource(baseService, filePath);
+        HttpIntrospectionResource introspectionResource = new HttpIntrospectionResource(httpService, filePath);
         byte[] payload = introspectionResource.getPayload();
         byte[] fileContent = new byte[0];
         try {
@@ -105,18 +105,18 @@ public class BaseServiceTest {
     @Test
     public void testGetNameOfIntrospectionResource() {
         BObject service = TestUtils.getNewServiceObject("hello");
-        BaseService baseService = new BaseService(service);
-        HttpIntrospectionResource introspectionResource = new HttpIntrospectionResource(baseService, "abc");
+        HttpService httpService = new HttpService(service);
+        HttpIntrospectionResource introspectionResource = new HttpIntrospectionResource(httpService, "abc");
         Assert.assertEquals(introspectionResource.getName(), "$get$openapi-doc-dygixywsw");
     }
 
     @Test
     public void testGetAbsoluteResourcePath() {
-        BaseService baseService = new BaseService(TestUtils.getNewServiceObject("hello"));
-        baseService.setBasePath("/basePath/");
-        BaseResource resource = new BaseResource(TestUtils.getNewMethodType(), baseService);
+        HttpService httpService = new HttpService(TestUtils.getNewServiceObject("hello"));
+        httpService.setBasePath("/basePath/");
+        HttpResource resource = new HttpResource(TestUtils.getNewMethodType(), httpService);
         resource.setPath("/abc/{xyz}");
-        baseService.setResources(Collections.singletonList(resource));
+        httpService.setResources(Collections.singletonList(resource));
         Assert.assertEquals(resource.getAbsoluteResourcePath(), "/basePath/abc/{xyz}");
     }
 }
